requirements.txt
jinja2
python-frontmatter
pyyaml
python-dateutil
jsonschema
beautifulsoup4
readability-lxml
markdownify
